Version files added x64 warning fixes
git-svn-id: http://svn.miranda-ng.org/main/trunk@433 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/ExtraIcons/ExtraIconGroup.cpp
@@ -122,8 +122,8 @@ int ExtraIconGroup::setIcon(int id, HANDLE hContact, void *icon)
 	}
 
 	ExtraIcon *current = getCurrentItem(hContact);
-	int currentPos = items.size();
-	int storePos = items.size();
+	int currentPos = (int)items.size();
+	int storePos = (int)items.size();
 	for (unsigned int i = 0; i < items.size(); ++i)
 	{
 		if (items[i]->getID() == id)
@@ -162,7 +162,7 @@ int ExtraIconGroup::setIcon(int id, HANDLE hContact, void *icon)
 
 			insideApply = true;
 
-			for (++storePos; storePos < items.size(); ++storePos)
+			for (++storePos; storePos < (int)items.size(); ++storePos)
 			{
 				items[storePos]->applyIcon(hContact);
